v2.4.9.9 -> v2.4.9.10
  - Alan Cox: continued merging
  - Mingming Cao: make msgrcv/shmat check the queue/segment ID's properly
  - Greg KH: USB serial init failure fix, Xircom serial converter driver
  - Neil Brown: nsfd/raid/md/lockd cleanups
  - Ingo Molnar: multipath RAID personality, raid xor update
  - Hugh Dickins/Marcelo Tosatti: swapin read-ahead race fix
  - Vojtech Pavlik: fix up some of the infrastructure for x86-64
  - Robert Love: AMD 761 AGP GART support
  - Jens Axboe: fix SCSI-generic queue handling race
  - me: be sane about page reference bits

+++ drivers/char/rio/riotable.c
@@ -445,7 +445,7 @@ struct rio_info *	p;
 	int Next = 0;
 	struct Map *MapP;
 	struct Host *HostP;
-	int oldspl;
+	long oldspl;
 
 	disable(oldspl);		/* strange but true! */
